// generated/src/app/api-models/case-caller.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CaseCaller {
  seqCallerId: number;
  seqCaseId: number;
  callerIsCustomerFlag: string;
  callerTypeCsCode: string;
  callerMasterId: string;
  callerMasterNumber: string;
  callerMasterSubNumber: string;
  contactMethodCsCode: string;
  relationCsCode: string;
  name: string;
  addr1: string;
  addr2: string;
  city: string;
  state: string;
  zipCode: string;
  phone1: string;
  phone1DesignCsCode: string;
  phone2: string;
  phone2DesignCsCode: string;
  userDefined1: string;
  userDefined2: string;
  userDefinedDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  country: string;
}